Generalize IdenAssertV01 code
Split verify IdenAssertV01 generic part related to identity into
verifySignedPacket function.

Appart from the split of functionality, this commit will break the usage of
signIdenAssertV01 because previously the generated signature was not following
the spec.  See changes in `src/protocols/login.test.js` and
`test/protocols-login.test.js` as an example of the breaking update.